add confirm: create new Post, delete post, cancel new post

+++ src/components/Post/Post.tsx
@@ -1,23 +1,33 @@
-import React from 'react';
+import React, {useMemo} from 'react';
 import styles from './Post.module.css'
 import {IPost} from "../../types";
 import {deletePost} from "../../redux/actionCreators";
 import {useAppDispatch} from "../../redux/hooks";
+import {UseConfirm} from "../../hooks/useConfirm";
 
-interface PostProps extends IPost {}
+interface PostProps extends IPost {
+}
 
 const Post: React.FC<PostProps> = ({id, title, author, text, date}) => {
-    const dateNormalize = new Date(date).toLocaleString();
+    const dateNormalize = useMemo(
+        () =>
+            new Date(date).toLocaleString(),
+        [date]
+    );
     const dispatch = useAppDispatch()
-    const handleDelete = () => {
-        dispatch(deletePost(id))
+    const {openConfirm} = UseConfirm()
+    const confirmHandleDelete = async () => {
+        const isConfirmed = await openConfirm('Удалить пост?')
+        if (isConfirmed) {
+            dispatch(deletePost(id))
+        }
     }
     return (
         <article className={styles.post}>
             <div className={styles.head}>
                 <h3>{title}</h3>
                 <span className="material-icons">edit</span>
-                <span onClick={handleDelete} className="material-icons">delete</span>
+                <span onClick={confirmHandleDelete} className="material-icons">delete</span>
             </div>
             <div className={styles.date}>Дата: {dateNormalize}</div>
             <div className={styles.author}>
